Add `p_global` to problem formulation in Python (#1234)
This allows users to define a CasADi variable `AcadosModel.p_global`
containing global parameters, i.e. parameters that can not change over
the horizon.
The code generation of external functions has been refactored using the
new `GenerateContext` class.
This class extracts all expressions that only depend on `p_global` from
the CasADi graph of the output expressions and replaces them with
"double memory pools".
This allows for an efficient precomputation of potentially expensive
terms which only depend on these parameters, e.g. spline coefficients,
when p_global are underlying data points.
However, since the double pools are part of the compilation unit of the
external functions, this implementation is inherently limited to global
parameters.
Implementing different parameters over the horizon for which expressions
are precomputed can be done by using an `AcadosMultiphaseOcp`
formulation and defining `p_global` as the concatenation of the
corresponding parameters of the phases.

Only supported for OCP solvers.
Updating these parameters can be done using
`AcadosOcpSolver.set_p_global(values)`
NOTE: this is only supported with CasADi beta release
https://github.com/casadi/casadi/releases/tag/nightly-se

---------

Co-authored-by: Joris Gillis <joris@yacoda.com>