Add option to reuse workspace of external CasADi functions (#1296)
This PR adds an option to reuse workspace of external CasADi functions
which can drastically reduce the memory footprint of OCP solvers.

- external functions: added option struct `external_function_opts`. In
there, the option `bool external_workspace` is available. If it is true,
the workspace for the external function has to be provided externally.
In the acados solver, this is then provided by the OCP-NLP solver or
integrator.
- In case of non parallelized OCP solvers, all external functions share
the same workspace, reducing the overall memory footprint of external
function workspaces to the maximum required by any function.
- In case of parallelization via OpenMP, the memory footprint is reduced
by sharing the workspace between external functions associated with the
same module of each shooting node.

- For now only float workspaces are shared, which typically make up the
largest share of the external function memory. Integer workspaces could
be added as well.
- For a reduced memory footprint with OpenMP, we could share the
workspaces among all modules associated with the same shooting node.
- We could use the same workspace reuse when compiling with
`-DACADOS_WITH_OPENMP=ON -DACADOS_NUM_THREADS=1` as when compiling
without OpenMP. This is in particular relevant for batch solvers.

- external function creation methods `external_function_*_create` got an
additional argument `external_function_opts *opts_`.
- The functions: `sim_solver_create`, `sim_solver_calculate_size`,
`sim_solver_assign` now need an additional argument `sim_in *in`, in
order to make the external functions available when the memory is
calculated and assigned.
- The functions: `sim_solver_create`, `sim_solver_calculate_size`,
`sim_solver_assign` now need an additional argument `sim_in *in`, in
order to make the external functions available when the memory is
calculated and assigned.
- The function `ocp_nlp_solver_create` now needs an additional argument
`ocp_nlp_in *nlp_in`, in order to make the external functions available
when the memory is calculated and assigned.